add essay.pdf for test
Thomas Malthus published his essay on Principles of Population in 1798. At the time, the descending "s" was in fashion and this reconstruction uses the prince-text-replace property to recreate the original printed book (Archive.org has a scanned copy). Also, the document uses ligatures and the Table of Contents is generated with a script.